Giant blocks done

+++ src/main/java/twilightforest/block/BlockTFGiantBlock.java
@@ -4,6 +4,7 @@ import net.minecraft.block.Block;
 import net.minecraft.block.material.PushReaction;
 import net.minecraft.block.BlockState;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.item.BlockItemUseContext;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
@@ -18,22 +19,22 @@ public class BlockTFGiantBlock extends Block {
 		super(props);
 	}
 
-//	@Override
-//	public boolean canPlaceBlockAt(World world, BlockPos pos) {
-//		for (BlockPos dPos : getVolume(pos)) {
-//			BlockState state = world.getBlockState(dPos);
-//			if (!state.getBlock().isReplaceable(world, dPos)) {
-//				return false;
-//			}
-//		}
-//		return super.canPlaceBlockAt(world, pos);
-//	}
+	@Nullable
+	@Override
+	public BlockState getStateForPlacement(BlockItemUseContext ctx) {
+		for (BlockPos dPos : getVolume(ctx.getPos())) {
+			if (!ctx.getWorld().getBlockState(dPos).isReplaceable(ctx)) {
+				return null;
+			}
+		}
+		return getDefaultState();
+	}
 
 	@Override
 	public void onBlockPlacedBy(World world, BlockPos pos, BlockState state, @Nullable LivingEntity placer, ItemStack stack) {
 		if (!world.isRemote) {
 			for (BlockPos dPos : getVolume(pos)) {
-				world.setBlockState(dPos, getDefaultState(), 2);
+				world.setBlockState(dPos, getDefaultState());
 			}
 		}
 	}
@@ -42,27 +43,25 @@ public class BlockTFGiantBlock extends Block {
 	@Deprecated
 	public void onReplaced(BlockState state, World world, BlockPos pos, BlockState newState, boolean isMoving) {
 		super.onReplaced(state, world, pos, newState, isMoving);
-		if (!this.isSelfDestructing && !canBlockStay(world, pos)) {
+		if (!this.isSelfDestructing && !isVolumeFilled(world, pos)) {
 			this.setGiantBlockToAir(world, pos);
 		}
 	}
 
 	private void setGiantBlockToAir(World world, BlockPos pos) {
-		// this flag is maybe not totally perfect
+		// prevent mutual infinite recursion
 		this.isSelfDestructing = true;
 
 		for (BlockPos iterPos : getVolume(pos)) {
-			if (!pos.equals(iterPos)) {
-				if (world.getBlockState(iterPos).getBlock() == this) {
-					world.destroyBlock(iterPos, false);
-				}
+			if (!pos.equals(iterPos) && world.getBlockState(iterPos).getBlock() == this) {
+				world.destroyBlock(iterPos, false);
 			}
 		}
 
 		this.isSelfDestructing = false;
 	}
 
-	private boolean canBlockStay(World world, BlockPos pos) {
+	private boolean isVolumeFilled(World world, BlockPos pos) {
 		for (BlockPos dPos : getVolume(pos)) {
 			if (world.getBlockState(dPos).getBlock() != this) {
 				return false;
@@ -75,10 +74,6 @@ public class BlockTFGiantBlock extends Block {
 	@Deprecated
 	public PushReaction getPushReaction(BlockState state) {
 		return PushReaction.BLOCK;
-	}
-
-	public static BlockPos roundCoords(BlockPos pos) {
-		return new BlockPos(pos.getX() & ~0b11, pos.getY() & ~0b11, pos.getZ() & ~0b11);
 	}
 
 	public static Iterable<BlockPos> getVolume(BlockPos pos) {
